fix: extract getInitials to other function
