+++ src/app/(authenticated)/profile/page.tsx
@@ -2,15 +2,7 @@ import { createClient } from '@/utils/supabase/server'
 import { redirect } from 'next/navigation'
 import { Card, CardContent, CardHeader, CardTitle, CardDescription } from "@/components/ui/card"
 import { Avatar, AvatarFallback } from "@/components/ui/avatar"
-
-function getInitials(email?: string): string {
-  if (!email) return 'U';
-  const parts = email.split('@')[0].split(/[\s._-]+/);
-  if (parts.length === 1) {
-    return parts[0].substring(0, 2).toUpperCase();
-  }
-  return (parts[0][0] + (parts[parts.length - 1][0] || '')).toUpperCase();
-}
+import { getInitials } from '@/utils/user'
 
 export default async function ProfilePage() {
   const supabase = await createClient()
